+++ flightsdb/query.go
@@ -9,7 +9,7 @@ import (
 func FindCheapestRoute(origin, destination *Airport) []*Flight {
 	routeTrace := make(map[*Airport]*Flight)
 	visited := make(map[*Airport]bool)
-	accPriceTable := make(map[*Airport]float32)
+	accPriceTable := make(map[*Airport]float64)
 	queue := &Queue{&QueueItem{data: origin, value: 0}}
 	accPriceTable[origin] = 0
 	heap.Init(queue)
